Struct clipboard_wayland::Clipboard

pub struct Clipboard { /* private fields */ }

Implementations

impl Clipboard

pub unsafe fn connect(display: *mut c_void) -> Clipboard

pub fn read(&self) -> Result<String, Box<dyn Error>>

pub fn read_primary(&self) -> Result<String, Box<dyn Error>>

pub fn write(&mut self, data: String) -> Result<(), Box<dyn Error>>

pub fn write_primary(&mut self, data: String) -> Result<(), Box<dyn Error>>

pub fn write_data<T: AsMimeTypes + Send + Sync + 'static>( &mut self, data: T, ) -> Result<(), Box<dyn Error>>

pub fn write_primary_data<T: AsMimeTypes + Send + Sync + 'static>( &mut self, data: T, ) -> Result<(), Box<dyn Error>>

pub fn read_data<T: AllowedMimeTypes + 'static>( &self, ) -> Result<T, Box<dyn Error>>

pub fn read_primary_data<T: AllowedMimeTypes + 'static>( &self, ) -> Result<T, Box<dyn Error>>

pub fn read_primary_raw( &self, allowed: Vec<String>, ) -> Result<(Vec<u8>, String), Box<dyn Error>>

pub fn read_raw( &self, allowed: Vec<String>, ) -> Result<(Vec<u8>, String), Box<dyn Error>>

pub fn init_dnd(&self, tx: DndSender)

pub fn start_dnd<D: AsMimeTypes + Send + 'static>( &self, internal: bool, source_surface: DndSurface, icon_surface: Option<Icon>, content: D, actions: DndAction, )

Start a DnD operation on the given surface with some data

pub fn end_dnd(&self)

End the current DnD operation, if there is one

pub fn register_dnd_destination( &self, surface: DndSurface, rectangles: Vec<DndDestinationRectangle>, )

Register a surface for receiving DnD offers Rectangles should be provided in order of decreasing priority. This method can be called multiple time for a single surface if the rectangles change.

pub fn set_action(&self, action: DndAction)

Set the final action after presenting the user with a choice

pub fn peek_offer<D: AllowedMimeTypes + 'static>( &self, mime_type: Option<Cow<'static, str>>, ) -> Result<D>

Peek at the contents of a DnD offer