clipboard_wayland

Trait AllowedMimeTypes

pub trait AllowedMimeTypes: TryFrom<(Vec<u8>, MimeType)> {
    // Required method
    fn allowed() -> Cow<'static, [MimeType]>;
}

Describes the mime types which are accepted.

Required Methods

fn allowed() -> Cow<'static, [MimeType]>

List allowed mime types for the type to convert from a byte slice.

Allowed mime types should be listed in order of decreasing preference, most preferred first.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.