Trait clipboard_wayland::AsMimeTypes

pub trait AsMimeTypes {
    // Required methods
    fn available(&self) -> Cow<'static, [MimeType]>;
    fn as_bytes(&self, mime_type: &MimeType) -> Option<Cow<'static, [u8]>>;
}

Can be converted to data with the available mime types.

Required Methods

fn available(&self) -> Cow<'static, [MimeType]>

List available mime types for this data to convert to a byte slice.

fn as_bytes(&self, mime_type: &MimeType) -> Option<Cow<'static, [u8]>>

Converts a type to a byte slice for the given mime type if possible.